Refactoring as per older PR comments - removing JmxSslRMIServerSocketFactory and moving jmx unit tests in separate package
